% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/comp.simulate.R
\name{comp.simulate}
\alias{comp.simulate}
\title{Simulate chronological distributions for subsets of a data table.}
\usage{
comp.simulate(data, probs = 1, weight = 1, comp.field = "group",
  comp.values = NULL, context.fields = c("SITE_C"), UoA = NULL,
  comp.fun = date.simulate, quant.list = c(0.025, 0.25, 0.5, 0.75, 0.975),
  start.date = 0, end.date = 2000, bin.width = 100, reps = 100,
  RoC = FALSE)
}
\arguments{
\item{data}{Data table (or object that can be coerced to one) with, minimally, two numeric columns called Start and End and a factor
column (or column that can be coerced to factor) on which to subset.}

\item{probs}{Numeric vector defining a null model from which to sample the dummy set. Will be recycled up to nrow(data), so passing a
single value results in a uniform null model. Defaults to 1.}

\item{weight}{Numeric vector: the weight to be applied to each row in `data` (and to its counterpart in the dummy set, where relevant),
or a constant weight to be applied to all. Defaults to 1.}

\item{comp.field}{Character: name of column to be used to subset 'data'. Defaults to "group".}

\item{comp.values}{Vector with length at least 2, specifying values of 'comp.field' to compare. Defaults to NULL, in which case all
unique values are compared.}

\item{context.fields}{Character vector specifying the column(s) in data which define the minimal stratigraphic entities to analyse.
Defaults to "SITE_C".}

\item{UoA}{Unit of Analysis: character vector of names of additional columns by which to group data when aggregating weights,
on top of those specified in context.field. For example, should different taxa be lumped together when analysing bone remains
from a table of contexts? Defaults to NULL.}

\item{comp.fun}{The function to apply to each subset of 'data'. Defaults to date.simulate.}

\item{quant.list}{Numeric vector of quantiles to be calculated in a summary table. Defaults to c(0.025,0.25,0.5,0.75,0.975).}

\item{start.date}{Numeric: the start of time period to be considered. Defaults to 0.}

\item{end.date}{Numeric: the end of time period to be considered. Defaults to 2000.}

\item{bin.width}{Numeric: the resolution of the analysis, in units of time. Defaults to 100.}

\item{reps}{Integer: the number of times the simulation will be run. Defaults to 100.}

\item{RoC}{Rate of Change. Logical: should rates of change between adjacent bins be calculated alongside the raw counts?}
}
\value{
A list with two named elements:
     "full" is a data table with at least five columns: 'rep.no', integer specifying simulation run; 'bin', character specifying
     chronological bin in terms of date range; 'bin.no' integer specifying number of bin, counting from earliest; then between one and
     four columns for each value specified by 'comp.values', depending on 'comp.fun' and 'RoC'.
     "summary" is a second long format data table with four named columns: 'bin', as above; 'V1', the relevant value for the given bin
     at a given quantile; 'quantile', the quantile at which V1 is calculated; 'id', character specifying which column from "full" V1 is
     based on: e.g "group1.count", "group1.dummy", "group4.RoC.count", "group5.RoC.dummy".
}
\description{
Applies date.simulate, dummy.simulate, or freq.simulate to two or more subsets of a data table, as defined by a single factor column.
}
\examples{
date.ranges <- data.table(unit=c(1, 2, 3, 4, 5), Start=c(450, 450, 600, 1000, 1200), End=c(700, 800, 650, 1200, 1525), frag.count=
     c(3, 6, 25, 1, 2), phase=c("a", "b", "b", "c", "c"))
x <- comp.simulate(date.ranges, comp.field="phase", comp.values=c("b", "c"), context.fields="unit", comp.fun=date.simulate, reps=200)
}

