% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/date.simulate.R
\name{date.simulate}
\alias{date.simulate}
\title{Simulate chronological frequency distribution.}
\usage{
date.simulate(data, weight = 1, context.fields = c("SITE_C"), UoA = NULL,
  start.date = 0, end.date = 2000, bin.width = 100, reps = 100,
  RoC = FALSE, summ = TRUE, ...)
}
\arguments{
\item{data}{Data table (or object that can be coerced to one) with, minimally, two numeric columns called Start and End.}

\item{weight}{Numeric vector: the weight to be applied to each row in `data`, or a constant weight to be applied to all.
Defaults to 1.}

\item{context.fields}{Character vector specifying the column(s) in data which define the minimal stratigraphic entities to analyse.
Defaults to "SITE_C".}

\item{UoA}{Unit of Analysis: character vector of names of additional columns by which to group data when aggregating weights,
on top of those specified in context.field. For example, should different taxa be lumped together when analysing bone remains
from a table of contexts? Defaults to NULL.}

\item{start.date}{Numeric: the start of time period to be considered. Defaults to 0.}

\item{end.date}{Numeric: the end of time period to be considered. Defaults to 2000.}

\item{bin.width}{Numeric: the resolution of the analysis, in units of time. Defaults to 100.}

\item{reps}{Integer: the number of times the simulation will be run. Defaults to 100.}

\item{RoC}{Rate of Change. Logical: should rates of change between adjacent bins be calculated alongside the raw counts?}

\item{summ}{Logical: should a summary results table be appended to the output? Defaults to TRUE. No real reason to change, except
when called as part of a more complex function. Nb. the quantiles to use when summarising can be specified using a numeric
vector called `quant.list`, which will be passed straight to the summarising function (`sim.summ`; see below) and defaults to
c(0.025,0.25,0.5,0.75,0.975).}
}
\value{
If summ=FALSE, a long-format data table with four or five named columns: 'rep.no', integer specifying simulation run; 'bin',
     character specifying chronological bin in terms of date range; 'bin.no' integer specifying number of bin, counting from earliest;
     'count', numeric giving the number of entities (or total weight) assigned to the given bin in the given simulation run; 'RoC'
     numeric giving the rate of change in count between this bin and the next (only present if RoC=TRUE).
     If summ=TRUE, a list with two named elements: "full" is as above; "summary" is a second long format data table with four named
     columns: 'bin', as above; 'V1', the relevant value for the given bin at a given quantile; 'quantile',
     the quantile at which V1 is calculated; 'id', character specifying which column from "full" V1 is based on - either
     "count" or "RoC".
}
\description{
Simulates chronological distributions from a table of entities with defined date ranges, based on assumption of uniform probability
      between limits.
}
\examples{
date.ranges <- data.table(unit=c(1, 2, 3, 4), Start=c(450, 450, 600, 1000), End=c(700, 800, 650, 1200), frags=c(3, 6, 25, 1))
x <- date.simulate(date.ranges, weight=date.ranges$frags, context.fields="unit", start.date=500, end.date=1500, bin.width=50, reps=200)
}

